Post refactoring - asset basic issuance * Refactor FindTypedEvent to use generics * Refactor the order of the FT issuance CLI args
