+++ integration-tests/asset/issue.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/CoreumFoundation/coreum/integration-tests/testing"
 	"github.com/CoreumFoundation/coreum/pkg/tx"
+	"github.com/CoreumFoundation/coreum/testutil/event"
 	assettypes "github.com/CoreumFoundation/coreum/x/asset/types"
 )
 
@@ -52,9 +53,7 @@ func TestIssueBasicFungibleToken(ctx context.Context, t testing.T, chain testing
 	)
 
 	require.NoError(t, err)
-	evt := testing.FindTypedEvent(t, &assettypes.EventFungibleTokenIssued{}, res.Events)
-	fungibleTokenIssuedEvt, ok := evt.(*assettypes.EventFungibleTokenIssued)
-	require.True(t, ok)
+	fungibleTokenIssuedEvt, err := event.FindTypedEvent(&assettypes.EventFungibleTokenIssued{}, res.Events)
 
 	require.NoError(t, err)
 	require.Equal(t, assettypes.EventFungibleTokenIssued{
