+++ x/asset/client/cli/tx.go
@@ -35,14 +35,14 @@ func GetTxCmd() *cobra.Command {
 // CmdTxIssueFungibleToken returns issue IssueFungibleToken cobra command.
 func CmdTxIssueFungibleToken() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "issue-ft [symbol] [description] [recipient_address] [initial_amount] --from [issuer]",
+		Use:   "issue-ft [symbol] [recipient_address] [initial_amount] [description] --from [issuer]",
 		Args:  cobra.ExactArgs(4),
 		Short: "Issue new fungible token",
 		Long: strings.TrimSpace(
 			fmt.Sprintf(`Issues new fungible token.
 
 Example:
-$ %s tx asset issue-ft BTC "BTC Token" [recipient_address] 100000 --from [issuer]
+$ %s tx asset issue-ft WBTC [recipient_address] 21000000 "Wrapped Bitcoin" --from [issuer]
 `,
 				version.AppName,
 			),
@@ -55,8 +55,7 @@ $ %s tx asset issue-ft BTC "BTC Token" [recipient_address] 100000 --from [issuer
 
 			issuer := clientCtx.GetFromAddress()
 			symbol := args[0]
-			description := args[1]
-			recipient := args[2]
+			recipient := args[1]
 			// if the recipient wasn't provided the signer is the recipient
 			if recipient != "" {
 				if _, err = sdk.AccAddressFromBech32(recipient); err != nil {
@@ -68,13 +67,15 @@ $ %s tx asset issue-ft BTC "BTC Token" [recipient_address] 100000 --from [issuer
 
 			// if the initial amount wasn't provided the amount is zero
 			initialAmount := sdk.ZeroInt()
-			if args[3] != "" {
+			if args[2] != "" {
 				var ok bool
-				initialAmount, ok = sdk.NewIntFromString(args[3])
+				initialAmount, ok = sdk.NewIntFromString(args[2])
 				if !ok {
 					return sdkerrors.Wrap(err, "invalid initial_amount")
 				}
 			}
+
+			description := args[3]
 
 			msg := &types.MsgIssueFungibleToken{
 				Issuer:        issuer.String(),
